map entitiy relationships

+++ src/main/java/com/kamil/TaskManagement/repository/TaskRepository.java
@@ -3,5 +3,6 @@ package com.kamil.TaskManagement.repository;
 import com.kamil.TaskManagement.model.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskRepository extends JpaRepository<Task, Long> {
+public interface TaskRepository extends JpaRepository<Task,Integer> {
+
 }
